+++ src/ArrayHelpers/DuplicateMapItemException.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+namespace Ingenerator\PHPUtils\ArrayHelpers;
+
+use RuntimeException;
+use UnexpectedValueException;
+
+class DuplicateMapItemException extends UnexpectedValueException
+{
+
+}